python/requirements.txt
Flask==3.0.0
python-dotenv==0.19.1
flask-security
flask-wtf
requests==2.31.0
gunicorn==21.2.0